requirements.txt
setuptools>=69.0.0
numpy>=1.26.0
sentence-transformers==2.2.2
transformers==4.34.0
torch>=2.2.0
faiss-cpu>=1.8.0
click==8.1.7
tqdm==4.66.1
fastapi==0.115.0
uvicorn==0.30.6
pydantic>=2.0.0
requests>=2.31.0
streamlit>=1.28.0
plotly>=5.17.0
networkx>=3.2.0
reportlab>=4.0.0
openai>=1.0.0
anthropic>=0.18.0
python-dotenv>=1.0.0